pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "2-owen-community-be"
version = "0.1.0"
description = "Community Backend API"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Owen", email = "owen@example.com" }
]
dependencies = [
    "fastapi==0.128.0",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.12.5",
    "pydantic-settings==2.6.1",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "requests==2.32.5",
    "starlette==0.50.0",
    "itsdangerous==2.2.0",
    "httpx==0.28.1",
    "email-validator>=2.0.0",
    "pytest>=8.0.0",
    "python-ulid==3.0.0",
    "bcrypt==5.0.0",
    "aiomysql",
    "Faker",
    "cryptography>=41.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["controllers*", "models*", "routers*", "utils*", "schemas*", "config.py", "main.py"]
